"Refactor: Enhance const-correctness and use modern C++ practices for better clarity, safety, and maintainability. Includes consistent `const` usage, `nullptr` over `NULL`, explicit type casting, and improved initializer syntax."

+++ src/openssl.h
@@ -237,7 +237,7 @@ struct CertStatusExData {
             return 0;
         }
 
-        if (BN_num_bytes(bn.get()) > sizeof(uint64_t)) {
+        if (static_cast<size_t>(BN_num_bytes(bn.get())) > sizeof(uint64_t)) {
             return 0;
         }
